+++ src/Handler/Profile.hs
@@ -1,12 +1,15 @@
-{-# LANGUAGE NoImplicitPrelude #-}
-{-# LANGUAGE OverloadedStrings #-}
-{-# LANGUAGE TemplateHaskell #-}
+{-# LANGUAGE NoImplicitPrelude     #-}
+{-# LANGUAGE OverloadedStrings     #-}
+{-# LANGUAGE TemplateHaskell       #-}
 {-# LANGUAGE MultiParamTypeClasses #-}
-{-# LANGUAGE TypeFamilies #-}
-module Handler.Profile where
+{-# LANGUAGE TypeFamilies          #-}
+
+module Handler.Profile
+    ( getProfileR )
+    where
 
 import Import
+import Handler.Home ( getNewHomeR )
 
--- TODO: profile page
 getProfileR :: Handler Html
-getProfileR = undefined
+getProfileR = getNewHomeR